ACM: fixed some memory cast warnings

+++ ArduCopter/commands.pde
@@ -35,7 +35,7 @@ static struct Location get_cmd_with_index(int i)
     }else{
         // we can load a command, we don't process it yet
         // read WP position
-        int32_t mem = (WP_START_BYTE) + (i * WP_SIZE);
+        uintptr_t mem = (WP_START_BYTE) + (i * WP_SIZE);
 
         temp.id = eeprom_read_byte((uint8_t*)mem);
 
@@ -84,7 +84,7 @@ static void set_cmd_with_index(struct Location temp, int i)
         temp.id = MAV_CMD_NAV_WAYPOINT;
     }
 
-    uint32_t mem = WP_START_BYTE + (i * WP_SIZE);
+    uintptr_t mem = WP_START_BYTE + (i * WP_SIZE);
 
     eeprom_write_byte((uint8_t *)   mem, temp.id);
